FNSR is enabled for everyone on PHP 8.1+
